Setup initial work creation transaction
This is the first chunk of an experimental Actor Stack replacement based on DRY
Transactions. The philosophy of this replacement is to ensure that
create/update/delete steps remain extensible but are granular (single
responsibility), stateless, reusable, and safe under error conditions.

We begin by introducing a `CreateWork` transaction that handles a few of the
basic steps involved in preparing a work for save:

  - `SetDefaultAdminSet`
  - `EnsureAdminSet` (fail if no AdminSet is present)
  - `EnsurePermissionTemplate` (fail if no PermissionTemplate is presesnt)
  - `SetModifiedDate`
  - `SetUploadedDate`
  - `SaveWork`

These steps are defined individually, and organized under the `CreateWork`
transacttion. They do not yet represent a full replacement of the behavior in
the `#create` branch of the default Actor Stack. Instead, they are intended to
grow in this direction.

Later work will provide the option for adopters to use a transaction like this
one in place of the Actor Stack.